Battle class moved to new module
